+++ src/test/java/com/iexec/sms/tee/challenge/TeeChallengeServiceTests.java
@@ -16,15 +16,20 @@
 
 package com.iexec.sms.tee.challenge;
 
+import com.iexec.commons.poco.task.TaskDescription;
+import com.iexec.sms.blockchain.IexecHubService;
 import com.iexec.sms.encryption.EncryptionService;
 import com.iexec.sms.secret.MeasuredSecretService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
-import org.mockito.MockitoAnnotations;
+import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 
+import java.security.GeneralSecurityException;
+import java.time.Instant;
 import java.util.Optional;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -32,17 +37,25 @@ import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.*;
 
 @DataJpaTest
+@ExtendWith(MockitoExtension.class)
 class TeeChallengeServiceTests {
 
     private static final String TASK_ID = "0x123";
     private static final String PLAIN_PRIVATE = "plainPrivate";
     private static final String ENC_PRIVATE = "encPrivate";
 
+    private final Instant finalDeadline = Instant.now().minusMillis(1000);
+
+    @Autowired
+    private EthereumCredentialsRepository ethereumCredentialsRepository;
     @Autowired
     private TeeChallengeRepository teeChallengeRepository;
 
     @Mock
     private EncryptionService encryptionService;
+
+    @Mock
+    private IexecHubService iexecHubService;
 
     @Mock
     private MeasuredSecretService teeChallengeMeasuredSecretService;
@@ -54,90 +67,136 @@ class TeeChallengeServiceTests {
 
     @BeforeEach
     void beforeEach() {
-        MockitoAnnotations.openMocks(this);
+        teeChallengeRepository.deleteAll();
         teeChallengeService = new TeeChallengeService(
                 teeChallengeRepository,
                 encryptionService,
+                iexecHubService,
                 teeChallengeMeasuredSecretService,
                 ethereumCredentialsMeasuredSecretService
         );
     }
 
-    private TeeChallenge getEncryptedTeeChallengeStub() throws Exception {
-        TeeChallenge teeChallenge = new TeeChallenge(TASK_ID);
+    private TeeChallenge getEncryptedTeeChallengeStub() throws GeneralSecurityException {
+        TeeChallenge teeChallenge = new TeeChallenge(TASK_ID, finalDeadline);
         teeChallenge.getCredentials().setEncryptedPrivateKey(ENC_PRIVATE);
         return teeChallenge;
     }
 
+    // region getOrCreate
     @Test
-    void shouldGetExistingChallengeWithoutDecryptingKeys() throws Exception {
+    void shouldGetExistingChallengeWithoutDecryptingKeys() throws GeneralSecurityException {
         TeeChallenge encryptedTeeChallengeStub = getEncryptedTeeChallengeStub();
         teeChallengeRepository.save(encryptedTeeChallengeStub);
 
         Optional<TeeChallenge> oTeeChallenge = teeChallengeService.getOrCreate(TASK_ID, false);
+
         assertThat(oTeeChallenge).isPresent();
         assertThat(oTeeChallenge.get().getCredentials().getPrivateKey()).isEqualTo(ENC_PRIVATE);
+        assertThat(teeChallengeRepository.count()).isOne();
+        assertThat(ethereumCredentialsRepository.count()).isOne();
         verify(encryptionService, never()).decrypt(anyString());
-
         verifyNoInteractions(teeChallengeMeasuredSecretService, ethereumCredentialsMeasuredSecretService);
     }
 
     @Test
-    void shouldGetExistingChallengeAndDecryptKeys() throws Exception {
+    void shouldGetExistingChallengeAndDecryptKeys() throws GeneralSecurityException {
         TeeChallenge encryptedTeeChallengeStub = getEncryptedTeeChallengeStub();
         teeChallengeRepository.save(encryptedTeeChallengeStub);
         when(encryptionService.decrypt(anyString())).thenReturn(PLAIN_PRIVATE);
 
         Optional<TeeChallenge> oTeeChallenge = teeChallengeService.getOrCreate(TASK_ID, true);
+
         assertThat(oTeeChallenge).isPresent();
         assertThat(oTeeChallenge.get().getCredentials().getPrivateKey()).isEqualTo(PLAIN_PRIVATE);
-        verify(encryptionService, times(1)).decrypt(anyString());
-
+        assertThat(teeChallengeRepository.count()).isOne();
+        assertThat(ethereumCredentialsRepository.count()).isOne();
+        verify(encryptionService).decrypt(anyString());
         verifyNoInteractions(teeChallengeMeasuredSecretService, ethereumCredentialsMeasuredSecretService);
     }
 
     @Test
     void shouldCreateNewChallengeWithoutDecryptingKeys() {
+        when(iexecHubService.getTaskDescription(TASK_ID))
+                .thenReturn(TaskDescription.builder().finalDeadline(finalDeadline.toEpochMilli()).build());
         when(encryptionService.encrypt(anyString())).thenReturn(ENC_PRIVATE);
 
         Optional<TeeChallenge> oTeeChallenge = teeChallengeService.getOrCreate(TASK_ID, false);
+
         assertThat(oTeeChallenge).isPresent();
         assertThat(oTeeChallenge.get().getCredentials().getPrivateKey()).isEqualTo(ENC_PRIVATE);
+        assertThat(teeChallengeRepository.count()).isOne();
+        assertThat(ethereumCredentialsRepository.count()).isOne();
         verify(encryptionService, never()).decrypt(anyString());
-
-        verify(teeChallengeMeasuredSecretService, times(1)).newlyAddedSecret();
-        verify(ethereumCredentialsMeasuredSecretService, times(1)).newlyAddedSecret();
+        verify(teeChallengeMeasuredSecretService).newlyAddedSecret();
+        verify(ethereumCredentialsMeasuredSecretService).newlyAddedSecret();
     }
 
     @Test
     void shouldCreateNewChallengeAndDecryptKeys() {
+        when(iexecHubService.getTaskDescription(TASK_ID))
+                .thenReturn(TaskDescription.builder().finalDeadline(finalDeadline.toEpochMilli()).build());
         when(encryptionService.encrypt(anyString())).thenReturn(ENC_PRIVATE);
         when(encryptionService.decrypt(anyString())).thenReturn(PLAIN_PRIVATE);
 
         Optional<TeeChallenge> oTeeChallenge = teeChallengeService.getOrCreate(TASK_ID, true);
+
         assertThat(oTeeChallenge).isPresent();
         assertThat(oTeeChallenge.get().getCredentials().getPrivateKey()).isEqualTo(PLAIN_PRIVATE);
-
-        verify(teeChallengeMeasuredSecretService, times(1)).newlyAddedSecret();
-        verify(ethereumCredentialsMeasuredSecretService, times(1)).newlyAddedSecret();
+        assertThat(teeChallengeRepository.count()).isOne();
+        assertThat(ethereumCredentialsRepository.count()).isOne();
+        verify(teeChallengeMeasuredSecretService).newlyAddedSecret();
+        verify(ethereumCredentialsMeasuredSecretService).newlyAddedSecret();
     }
+    // endregion
 
+    // region encryptChallengeKeys
     @Test
-    void shouldEncryptChallengeKeys() throws Exception {
-        TeeChallenge teeChallenge = new TeeChallenge(TASK_ID);
+    void shouldEncryptChallengeKeys() throws GeneralSecurityException {
+        TeeChallenge teeChallenge = new TeeChallenge(TASK_ID, finalDeadline);
         when(encryptionService.encrypt(anyString())).thenReturn(ENC_PRIVATE);
         teeChallengeService.encryptChallengeKeys(teeChallenge);
 
         assertThat(teeChallenge.getCredentials().getPrivateKey()).isEqualTo(ENC_PRIVATE);
     }
+    // endregion
 
+    // region decryptChallengeKeys
     @Test
-    void shouldDecryptChallengeKeys() throws Exception {
-        TeeChallenge teeChallenge = new TeeChallenge(TASK_ID);
-        teeChallenge.getCredentials().setEncryptedPrivateKey(ENC_PRIVATE);
+    void shouldDecryptChallengeKeys() {
+        TeeChallenge teeChallenge = TeeChallenge.builder()
+                .taskId(TASK_ID)
+                .credentials(new EthereumCredentials("id", "pk", true, "address"))
+                .build();
         when(encryptionService.decrypt(anyString())).thenReturn(PLAIN_PRIVATE);
 
         teeChallengeService.decryptChallengeKeys(teeChallenge);
         assertThat(teeChallenge.getCredentials().getPrivateKey()).isEqualTo(PLAIN_PRIVATE);
     }
+    // endregion
+
+    // region cleanExpiredChallenge
+    @Test
+    void shouldPurgeExpiredChallenge() throws GeneralSecurityException {
+        TeeChallenge encryptedTeeChallengeStub = getEncryptedTeeChallengeStub();
+        teeChallengeRepository.save(encryptedTeeChallengeStub);
+
+        teeChallengeService.cleanExpiredTasksTeeChallenges();
+
+        assertThat(teeChallengeRepository.count()).isZero();
+        assertThat(ethereumCredentialsRepository.count()).isZero();
+    }
+
+    @Test
+    void shouldNotPurgeExpiredChallenge() throws GeneralSecurityException {
+        TeeChallenge teeChallenge = new TeeChallenge(TASK_ID, null);
+        teeChallenge.getCredentials().setEncryptedPrivateKey(ENC_PRIVATE);
+        teeChallengeRepository.save(teeChallenge);
+
+        teeChallengeService.cleanExpiredTasksTeeChallenges();
+
+        assertThat(teeChallengeRepository.count()).isOne();
+        assertThat(ethereumCredentialsRepository.count()).isOne();
+    }
+    // endregion
 }