+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file.
 - Accept scheduler default result-proxy as a web2 secret to fallback on it when no proxy
   is specified in deal parameters. (#273)
 - Configure the SMS at startup to generate Scone sessions in Hardware or MAA mode. (#275)
+- Add configurable cron job to delete expired tasks TEE challenges and Ethereum credentials. (#278)
 
 ### Quality